Move CRDT tests to their own test file
