+++ test/log.spec.js
@@ -1799,206 +1799,5 @@ Object.keys(testAPIs).forEach((IPFS) => {
         assert.strictEqual(log4.clock.id, testIdentity4.publicKey)
       })
     })
-
-    describe('is a CRDT', () => {
-      let log1, log2, log3
-
-      beforeEach(async () => {
-        log1 = new Log(ipfs, testACL, testIdentity, 'X')
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        log3 = new Log(ipfs, testACL, testIdentity3, 'X')
-      })
-
-      it('join is associative', async () => {
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-        await log3.append('helloC1')
-        await log3.append('helloC2')
-
-        // a + (b + c)
-        await log2.join(log3)
-        await log1.join(log2)
-
-        const res1 = log1.values.slice()// .map((e) => e.hash).join(",")
-
-        log1 = new Log(ipfs, testACL, testIdentity, 'X')
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        log3 = new Log(ipfs, testACL, testIdentity3, 'X')
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-        await log3.append('helloC1')
-        await log3.append('helloC2')
-
-        // (a + b) + c
-        await log1.join(log2)
-        await log3.join(log1)
-
-        const res2 = log3.values.slice()// .map((e) => e.hash).join(",")
-
-        // associativity: a + (b + c) == (a + b) + c
-        const len = 6// (46 + 1) * 6- 1 // 46 == ipfs hash, +1 == .join(","), * 4 == number of items, -1 == last item doesn't get a ',' from .join
-        assert.strictEqual(res1.length, len)
-        assert.strictEqual(res2.length, len)
-        assert.deepStrictEqual(res1, res2)
-      })
-
-      it('join is commutative', async () => {
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-
-        // b + a
-        await log2.join(log1)
-        const res1 = log2.values.slice()// .map((e) => e.hash).join(",")
-
-        log1 = new Log(ipfs, testACL, testIdentity, 'X')
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-
-        // a + b
-        await log1.join(log2)
-        const res2 = log1.values.slice()// .map((e) => e.hash).join(",")
-
-        // commutativity: a + b == b + a
-        // const len = (46 + 1) * 4 - 1 // 46 == ipfs hash length, +1 == .join(","), * 4 == number of items, -1 == last item doesn't get a ',' from .join
-        assert.strictEqual(res1.length, 4)
-        assert.strictEqual(res2.length, 4)
-        assert.deepStrictEqual(res1, res2)
-      })
-
-      it('multiple joins are commutative', async () => {
-        // b + a == a + b
-        log1 = new Log(ipfs, testACL, testIdentity, 'X')
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-        await log2.join(log1)
-        const resA1 = log2.toString()
-
-        log1 = new Log(ipfs, testACL, testIdentity, 'X')
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-        await log1.join(log2)
-        const resA2 = log1.toString()
-
-        assert.strictEqual(resA1, resA2)
-
-        // a + b == b + a
-        log1 = new Log(ipfs, testACL, testIdentity, 'X')
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-        await log1.join(log2)
-        const resB1 = log1.toString()
-
-        log1 = new Log(ipfs, testACL, testIdentity, 'X')
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-        await log2.join(log1)
-        const resB2 = log2.toString()
-
-        assert.strictEqual(resB1, resB2)
-
-        // a + c == c + a
-        log1 = new Log(ipfs, testACL, testIdentity, 'A')
-        log3 = new Log(ipfs, testACL, testIdentity3, 'A')
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log3.append('helloC1')
-        await log3.append('helloC2')
-        await log3.join(log1)
-        const resC1 = log3.toString()
-
-        log1 = new Log(ipfs, testACL, testIdentity, 'X')
-        log3 = new Log(ipfs, testACL, testIdentity3, 'X')
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log3.append('helloC1')
-        await log3.append('helloC2')
-        await log1.join(log3)
-        const resC2 = log1.toString()
-
-        assert.strictEqual(resC1, resC2)
-
-        // c + b == b + c
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        log3 = new Log(ipfs, testACL, testIdentity3, 'X')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-        await log3.append('helloC1')
-        await log3.append('helloC2')
-        await log3.join(log2)
-        const resD1 = log3.toString()
-
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        log3 = new Log(ipfs, testACL, testIdentity3, 'X')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-        await log3.append('helloC1')
-        await log3.append('helloC2')
-        await log2.join(log3)
-        const resD2 = log2.toString()
-
-        assert.strictEqual(resD1, resD2)
-
-        // a + b + c == c + b + a
-        log1 = new Log(ipfs, testACL, testIdentity, 'X')
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        log3 = new Log(ipfs, testACL, testIdentity3, 'X')
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-        await log3.append('helloC1')
-        await log3.append('helloC2')
-        await log1.join(log2)
-        await log1.join(log3)
-        const logLeft = log1.toString()
-
-        log1 = new Log(ipfs, testACL, testIdentity, 'X')
-        log2 = new Log(ipfs, testACL, testIdentity2, 'X')
-        log3 = new Log(ipfs, testACL, testIdentity3, 'X')
-        await log1.append('helloA1')
-        await log1.append('helloA2')
-        await log2.append('helloB1')
-        await log2.append('helloB2')
-        await log3.append('helloC1')
-        await log3.append('helloC2')
-        await log3.join(log2)
-        await log3.join(log1)
-        const logRight = log3.toString()
-
-        assert.strictEqual(logLeft, logRight)
-      })
-
-      it('join is idempotent', async () => {
-        let logA = new Log(ipfs, testACL, testIdentity, 'X')
-        await logA.append('helloA1')
-        await logA.append('helloA2')
-        await logA.append('helloA3')
-
-        // idempotence: a + a = a
-        await logA.join(logA)
-        assert.strictEqual(logA.length, 3)
-      })
-    })
   })
 })
